pyproject.toml
[project]
name = "coding-agent"
version = "0.1.0"
description = "An AI-powered CLI coding assistant agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.66.0",
    "groq>=1.1.0",
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[project.scripts]
koda = "agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]
